+++ src/Exception/FileNotExistsException.php
@@ -17,6 +17,7 @@ declare(strict_types=1);
 namespace Tools\Exception;
 
 use Tools\FileException;
+use Tools\Filesystem;
 
 /**
  * "File or directory does not exist" exception.
@@ -33,7 +34,7 @@ class FileNotExistsException extends FileException
     public function __construct(?string $message = null, int $code = 0, ?\Throwable $previous = null, ?string $path = null)
     {
         if (!$message) {
-            $message = $path ? sprintf('Filename `%s` does not exist', rtr($path)) : 'Filename does not exist';
+            $message = $path ? sprintf('Filename `%s` does not exist', (new Filesystem())->rtr($path)) : 'Filename does not exist';
         }
         parent::__construct($message, $code, $previous, $path);
     }