+++ src/FileArray.php
@@ -17,6 +17,7 @@ declare(strict_types=1);
 namespace Tools;
 
 use Tools\Exceptionist;
+use Tools\Filesystem;
 
 /**
  * This class allows you to read and write arrays using text files
@@ -159,6 +160,6 @@ class FileArray
      */
     public function write(): bool
     {
-        return create_file($this->filename, serialize($this->data));
+        return (new Filesystem())->createFile($this->filename, serialize($this->data));
     }
 }