+++ src/Filesystem.php
@@ -11,28 +11,32 @@ declare(strict_types=1);
  * @copyright   Copyright (c) Mirko Pagliai
  * @link        https://github.com/mirko-pagliai/php-tools
  * @license     https://opensource.org/licenses/mit-license.php MIT License
+ * @since       1.4.4
  */
 
+namespace Tools;
+
 use Symfony\Component\Filesystem\Exception\IOException;
-use Symfony\Component\Filesystem\Filesystem;
+use Symfony\Component\Filesystem\Filesystem as BaseFilesystem;
 use Symfony\Component\Finder\Exception\DirectoryNotFoundException;
 use Symfony\Component\Finder\Finder;
 use Tools\Exceptionist;
 
-if (!function_exists('add_slash_term')) {
+/**
+ * Provides basic utility to manipulate the file system.
+ */
+class Filesystem extends BaseFilesystem
+{
     /**
      * Adds the slash term to a path, if it doesn't have one
      * @param string $path Path
      * @return string Path with the slash term
-     * @since 1.2.6
      */
-    function add_slash_term(string $path): string
+    public function addSlashTerm(string $path): string
     {
-        return is_slash_term($path) ? $path : $path . DS;
+        return $this->isSlashTerm($path) ? $path : $path . DS;
     }
-}
 
-if (!function_exists('create_file')) {
     /**
      * Creates a file. Alias for `mkdir()` and `file_put_contents()`.
      *
@@ -43,15 +47,13 @@ if (!function_exists('create_file')) {
      * @param int $dirMode Mode for the directory, if it does not exist
      * @param bool $ignoreErrors With `true`, errors will be ignored
      * @return bool
-     * @since 1.1.7
      * @throws \Symfony\Component\Filesystem\Exception\IOException
      */
-    function create_file(string $filename, $data = null, int $dirMode = 0777, bool $ignoreErrors = false): bool
+    public function createFile(string $filename, $data = null, int $dirMode = 0777, bool $ignoreErrors = false): bool
     {
         try {
-            $filesystem = new Filesystem();
-            $filesystem->mkdir(dirname($filename), $dirMode);
-            $filesystem->dumpFile($filename, $data);
+            $this->mkdir(dirname($filename), $dirMode);
+            $this->dumpFile($filename, $data);
 
             return true;
         } catch (IOException $e) {
@@ -62,9 +64,7 @@ if (!function_exists('create_file')) {
             return false;
         }
     }
-}
 
-if (!function_exists('create_tmp_file')) {
     /**
      * Creates a tenporary file. Alias for `tempnam()` and `file_put_contents()`.
      *
@@ -77,18 +77,15 @@ if (!function_exists('create_tmp_file')) {
      *  be created
      * @param string|null $prefix The prefix of the generated temporary filename
      * @return string Path of temporary filename
-     * @since 1.1.7
      */
-    function create_tmp_file($data = null, ?string $dir = null, ?string $prefix = 'tmp'): string
+    public function createTmpFile($data = null, ?string $dir = null, ?string $prefix = 'tmp'): string
     {
         $filename = @tempnam($dir ?: (defined('TMP') ? TMP : sys_get_temp_dir()), $prefix);
-        create_file($filename, $data);
+        $this->createFile($filename, $data);
 
         return $filename;
     }
-}
 
-if (!function_exists('dir_tree')) {
     /**
      * Returns an array of nested directories and files in each directory
      * @param string $path The directory path to build the tree from
@@ -96,10 +93,9 @@ if (!function_exists('dir_tree')) {
      *  to exclude or boolean true to not grab dot files/folders
      * @param bool $ignoreErrors With `true`, errors will be ignored
      * @return array Array of nested directories and files in each directory
-     * @since 1.0.7
      * @throws \Symfony\Component\Finder\Exception\DirectoryNotFoundException
      */
-    function dir_tree(string $path, $exceptions = false, bool $ignoreErrors = false): array
+    public function getDirTree(string $path, $exceptions = false, bool $ignoreErrors = false): array
     {
         $path = $path === DS ? DS : rtrim($path, DS);
         $finder = new Finder();
@@ -139,39 +135,7 @@ if (!function_exists('dir_tree')) {
             return [[], []];
         }
     }
-}
 
-if (!function_exists('fileperms_as_octal')) {
-    /**
-     * Gets permissions for the given file.
-     *
-     * Unlike the `fileperms()` function provided by PHP, this function returns
-     *  the permissions as four-chars string
-     * @link http://php.net/manual/en/function.fileperms.php
-     * @param string $filename Path to the file
-     * @return string Permissions as four-chars string
-     * @since 1.2.0
-     */
-    function fileperms_as_octal(string $filename): string
-    {
-        return (string)substr(sprintf('%o', fileperms($filename)), -4);
-    }
-}
-
-if (!function_exists('fileperms_to_string')) {
-    /**
-     * Returns permissions from octal value (`0755`) to string (`'0755'`)
-     * @param int|string $perms Permissions as octal value
-     * @return string Permissions as four-chars string
-     * @since 1.2.0
-     */
-    function fileperms_to_string($perms): string
-    {
-        return is_string($perms) ? $perms : sprintf('%04o', $perms);
-    }
-}
-
-if (!function_exists('get_extension')) {
     /**
      * Gets the extension from a filename.
      *
@@ -180,9 +144,8 @@ if (!function_exists('get_extension')) {
      *  several parts (eg, `sql.gz`).
      * @param string $filename Filename
      * @return string|null
-     * @since 1.0.2
      */
-    function get_extension(string $filename): ?string
+    public function getExtension(string $filename): ?string
     {
         //Gets the basename and, if the filename is an url, removes query string
         //  and fragments (#)
@@ -200,22 +163,17 @@ if (!function_exists('get_extension')) {
 
         return $pos === false ? null : strtolower(substr($filename, $pos + 1));
     }
-}
 
-if (!function_exists('is_slash_term')) {
     /**
      * Checks if a path ends in a slash (i.e. is slash-terminated)
      * @param string $path Path
      * @return bool
-     * @since 1.0.3
      */
-    function is_slash_term(string $path): bool
+    public function isSlashTerm(string $path): bool
     {
         return in_array($path[strlen($path) - 1], ['/', '\\']);
     }
-}
 
-if (!function_exists('is_writable_resursive')) {
     /**
      * Tells whether a directory and its subdirectories are writable.
      *
@@ -224,13 +182,12 @@ if (!function_exists('is_writable_resursive')) {
      * @param bool $checkOnlyDir If `true`, also checks for all files
      * @param bool $ignoreErrors With `true`, errors will be ignored
      * @return bool
-     * @since 1.0.7
      * @throws \Symfony\Component\Finder\Exception\DirectoryNotFoundException
      */
-    function is_writable_resursive(string $dirname, bool $checkOnlyDir = true, bool $ignoreErrors = false): bool
+    public function isWritableResursive(string $dirname, bool $checkOnlyDir = true, bool $ignoreErrors = false): bool
     {
         try {
-            [$directories, $files] = dir_tree($dirname);
+            [$directories, $files] = $this->getDirTree($dirname);
             $items = $checkOnlyDir ? $directories : array_merge($directories, $files);
 
             if (!in_array($dirname, $items)) {
@@ -252,36 +209,29 @@ if (!function_exists('is_writable_resursive')) {
             return false;
         }
     }
-}
 
-if (!function_exists('rmdir_recursive')) {
     /**
      * Removes the directory itself and all its contents, including
      *  subdirectories and files.
      *
      * To remove only files contained in a directory and its sub-directories,
-     *  leaving the directories unaltered, use the `unlink_recursive()`
-     *  function instead.
+     *  leaving the directories unaltered, use the `unlinkRecursive()`
+     *  method instead.
      * @param string $dirname Path to the directory
      * @return bool
-     * @see unlink_recursive()
-     * @since 1.0.6
+     * @see \Tools\Filesystem::unlinkRecursive()
      * @throws \Symfony\Component\Filesystem\Exception\IOException
      */
-    function rmdir_recursive(string $dirname): bool
+    public function rmdirRecursive(string $dirname): bool
     {
         if (!is_dir($dirname)) {
             return false;
         }
-
-        $filesystem = new Filesystem();
-        $filesystem->remove($dirname);
+        $this->remove($dirname);
 
         return true;
     }
-}
 
-if (!function_exists('rtr')) {
     /**
      * Returns a path relative to the root.
      *
@@ -291,7 +241,7 @@ if (!function_exists('rtr')) {
      * @return string Relative path
      * @throws \Exception
      */
-    function rtr(string $path): string
+    public function rtr(string $path): string
     {
         $root = getenv('ROOT');
         if (!$root) {
@@ -299,39 +249,34 @@ if (!function_exists('rtr')) {
             $root = ROOT;
         }
 
-        $filesystem = new Filesystem();
-        if ($filesystem->isAbsolutePath($path) && string_starts_with($path, $root)) {
-            $path = $filesystem->makePathRelative($path, $root);
+        if ($this->isAbsolutePath($path) && string_starts_with($path, $root)) {
+            $path = $this->makePathRelative($path, $root);
         }
 
         return rtrim($path, '/');
     }
-}
 
-if (!function_exists('unlink_recursive')) {
     /**
      * Recursively removes all the files contained in a directory and within its
      *  sub-directories. This function only removes the files, leaving the
      *  directories unaltered.
      *
      * To remove the directory itself and all its contents, use the
-     *  `rmdir_recursive()` function instead.
+     *  `rmdirRecursive()` method instead.
      * @param string $dirname The directory path
      * @param array|bool $exceptions Either an array of files to exclude
      *  or boolean true to not grab dot files
      * @param bool $ignoreErrors With `true`, errors will be ignored
      * @return bool
-     * @see rmdir_recursive()
-     * @since 1.0.7
+     * @see \Tools\Filesystem::rmdirRecursive()
      * @throws \Symfony\Component\Filesystem\Exception\IOException
      * @throws \Symfony\Component\Finder\Exception\DirectoryNotFoundException
      */
-    function unlink_recursive(string $dirname, $exceptions = false, bool $ignoreErrors = false): bool
+    public function unlinkRecursive(string $dirname, $exceptions = false, bool $ignoreErrors = false): bool
     {
         try {
-            [, $files] = dir_tree($dirname, $exceptions);
-            $filesystem = new Filesystem();
-            $filesystem->remove($files);
+            [, $files] = $this->getDirTree($dirname, $exceptions);
+            $this->remove($files);
 
             return true;
         } catch (IOException | DirectoryNotFoundException $e) {