+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # 1.x branch
 ## 1.5 branch
 ### 1.5.0
+* all filesystem global function have been removed, use `Filesystem` class instead;
+* `TestTrait::assertFilePerms()` has been removed. Use instead
+    `assertFileIsReadable()`/`assertFileIsWritable()`/
+    `assertDirectoryIsReadable()`/`assertDirectoryIsWritable()`;
 * the `ReflectionTrait` has been moved on `Tools\TestSuite` namespace;
 * changed the order of arguments for the `TestTrait::assertException()` method.
     The callable is now the first argument;
@@ -8,6 +12,12 @@
 
 ## 1.4 branch
 ### 1.4.4
+* added `Filesystem` class, all filesystem global function are now deprecated;
+* `TestTrait::assertFilePerms()` is deprecated. Use instead
+    `assertFileIsReadable()`/`assertFileIsWritable()`/
+    `assertDirectoryIsReadable()`/`assertDirectoryIsWritable()`;
+* `fileperms_as_octal()` and `fileperms_to_string()` global functions are now
+    deprecated and will be removed in a future release;
 * some functions have been moved to `array_functions.php` file.
 
 ### 1.4.3